+++ src/main/java/carpet/utils/CarpetRulePrinter.java
@@ -5,8 +5,6 @@ import joptsimple.OptionParser;
 import joptsimple.OptionSet;
 import joptsimple.OptionSpec;
 import joptsimple.util.PathConverter;
-//import net.fabricmc.api.DedicatedServerModInitializer;
-//import net.fabricmc.loader.api.FabricLoader;
 
 import java.io.IOException;
 import java.io.PrintStream;
@@ -17,7 +15,9 @@ import java.util.Arrays;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.thinkingstudio.bedsheet.util.NeoHooks;
+
+import org.thinkingstudio.bedsheet.loader.FoxifiedLoader;
+import org.thinkingstudio.bedsheet.loader.entrypoint.DedicatedServerModInitializer;
 
 /**
  * Provides a command line interface to generate a dump with all rules
@@ -25,11 +25,12 @@ import org.thinkingstudio.bedsheet.util.NeoHooks;
  * category filter
  *
  */
-public class CarpetRulePrinter /*implements DedicatedServerModInitializer*/ {
-    public static void onInitializeServer() {
+public class CarpetRulePrinter implements DedicatedServerModInitializer {
+    @Override
+    public void onInitializeServer() {
         // When launching, we use the "--" separator to prevent the game rejecting to launch because of unknown options
         // Clear it in case it's present given else our option parser would also ignore them!
-        String[] args = Arrays.stream(NeoHooks.getLaunchArguments(true)).filter(opt -> !opt.equals("--")).toArray(String[]::new);
+        String[] args = Arrays.stream(FoxifiedLoader.getLaunchArguments(true)).filter(opt -> !opt.equals("--")).toArray(String[]::new);
 
         // Prepare an OptionParser for our parameters
         OptionParser parser = new OptionParser();