+++ src/main/java/org/thinkingstudio/bedsheet/BedSheetModEntrypoint.java
@@ -2,18 +2,23 @@ package org.thinkingstudio.bedsheet;
 
 import carpet.CarpetServer;
 import carpet.utils.CarpetRulePrinter;
+import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
+import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.loading.FMLLoader;
 import net.neoforged.neoforge.common.NeoForge;
+import org.thinkingstudio.bedsheet.loader.entrypoint.DedicatedServerModInitializer;
+import org.thinkingstudio.bedsheet.loader.entrypoint.EntrypointHandler;
 
-@Mod(BedSheetModReference.MODID)
+@Mod(value = BedSheetModReference.MODID, dist = Dist.DEDICATED_SERVER)
 public class BedSheetModEntrypoint {
-    public BedSheetModEntrypoint(IEventBus modEventBus) {
-        CarpetServer.onGameStarted();
-        BedSheetModEvents.registerEvents(modEventBus, NeoForge.EVENT_BUS);
+    public BedSheetModEntrypoint(ModContainer modContainer, IEventBus modEventBus) {
+        EntrypointHandler.init(modEventBus);
         if (FMLLoader.getDist().isDedicatedServer()) {
-            CarpetRulePrinter.onInitializeServer();
+            CarpetServer.onGameStarted();
+            BedSheetModEvents.registerEvents(modEventBus, NeoForge.EVENT_BUS);
+            modContainer.registerExtensionPoint(DedicatedServerModInitializer.class, new CarpetRulePrinter());
         }
     }
 }