+++ src/main/java/org/thinkingstudio/bedsheet/loader/entrypoint/DedicatedServerModInitializer.java
@@ -1,0 +1,7 @@
+package org.thinkingstudio.bedsheet.loader.entrypoint;
+
+import net.neoforged.fml.IExtensionPoint;
+
+public interface DedicatedServerModInitializer extends IExtensionPoint {
+    void onInitializeServer();
+}