+++ src/main/java/org/thinkingstudio/bedsheet/loader/entrypoint/ModInitializer.java
@@ -1,0 +1,7 @@
+package org.thinkingstudio.bedsheet.loader.entrypoint;
+
+import net.neoforged.fml.IExtensionPoint;
+
+public interface ModInitializer extends IExtensionPoint {
+    void onInitialize();
+}