+++ CHANGELOG.md
@@ -1,4 +1,2 @@
 ## Changelog
-- rework `customSticky` mixin
-- fix conflict with worldedit
-- update NeoForge version to `21.4.123`
+- change: `NeoHooks` -> `FoxifiedLoader`